+++ EcommerceApp.Web.ViewModels/FavoriteProduct/DeleteFavoriteViewModel.cs
@@ -13,8 +13,8 @@
         public string ProductName { get; set; } = null!;
 
         [Required]
-        [Range(typeof(int), ProductPriceMinValue, ProductPriceMaxValue)]
-        public int ProductPrice { get; set; }
+        [Range(typeof(decimal), ProductPriceMinValue, ProductPriceMaxValue)]
+        public decimal ProductPrice { get; set; }
 
         [Required]
         [MinLength(ProductDescriptionMinLength)]
